+++ packages/@livestore/livestore/src/store/create-store.ts
@@ -1,16 +1,17 @@
-import type {
-  Adapter,
-  BootStatus,
-  ClientSession,
-  ClientSessionDevtoolsChannel,
-  ClientSessionSyncProcessorSimulationParams,
-  IntentionalShutdownCause,
-  MigrationsReport,
+import {
+  type Adapter,
+  type BootStatus,
+  type ClientSession,
+  type ClientSessionDevtoolsChannel,
+  type ClientSessionSyncProcessorSimulationParams,
+  type IntentionalShutdownCause,
+  type MigrationsReport,
+  provideOtel,
+  type SyncError,
+  UnexpectedError,
 } from '@livestore/common'
-import { provideOtel, UnexpectedError } from '@livestore/common'
 import type { LiveStoreSchema } from '@livestore/common/schema'
 import { isDevEnv, LS_DEV } from '@livestore/utils'
-import type { Cause, Schema } from '@livestore/utils/effect'
 import {
   Context,
   Deferred,
@@ -24,6 +25,7 @@ import {
   OtelTracer,
   Queue,
   Runtime,
+  type Schema,
   Scope,
   TaskTracing,
 } from '@livestore/utils/effect'
@@ -215,9 +217,9 @@ export const createStore = <TSchema extends LiveStoreSchema = LiveStoreSchema, T
 
       const runtime = yield* Effect.runtime<Scope.Scope>()
 
-      const shutdown = (cause: Cause.Cause<UnexpectedError | IntentionalShutdownCause>) =>
+      const shutdown = (exit: Exit.Exit<IntentionalShutdownCause, UnexpectedError | SyncError>) =>
         Effect.gen(function* () {
-          yield* Scope.close(lifetimeScope, Exit.failCause(cause)).pipe(
+          yield* Scope.close(lifetimeScope, exit).pipe(
             Effect.logWarnIfTakesLongerThan({ label: '@livestore/livestore:shutdown', duration: 500 }),
             Effect.timeout(1000),
             Effect.catchTag('TimeoutException', () =>
@@ -226,7 +228,7 @@ export const createStore = <TSchema extends LiveStoreSchema = LiveStoreSchema, T
           )
 
           if (shutdownDeferred) {
-            yield* Deferred.failCause(shutdownDeferred, cause)
+            yield* Deferred.done(shutdownDeferred, exit)
           }
 
           yield* Effect.logDebug('LiveStore shutdown complete')