+++ packages/@livestore/sqlite-wasm/src/node/NodeFS.ts
@@ -19,10 +19,12 @@ export class NodeFS extends FacadeVFS {
   // biome-ignore lint/correctness/noUnusedPrivateClassMembers: for debugging
   private lastError: Error | null = null
   private readonly directory: string
-  constructor(name: string, sqlite3: WaSqlite.SQLiteAPI, directory: string) {
+  private readonly onShutdown: () => void
+  constructor(name: string, sqlite3: WaSqlite.SQLiteAPI, directory: string, onShutdown: () => void) {
     super(name, sqlite3)
 
     this.directory = directory
+    this.onShutdown = onShutdown
   }
 
   getFilename(fileId: number): string {
@@ -102,6 +104,8 @@ export class NodeFS extends FacadeVFS {
 
   jClose(fileId: number): number {
     try {
+      this.onShutdown()
+
       const file = this.mapIdToFile.get(fileId)
       if (!file) return VFS.SQLITE_OK
 
@@ -163,6 +167,8 @@ export class NodeFS extends FacadeVFS {
 
   jDelete(zName: string, _syncDir: number): number {
     try {
+      this.onShutdown()
+
       const pathname = path.resolve(this.directory, zName)
       fs.unlinkSync(pathname)
       return VFS.SQLITE_OK
@@ -185,6 +191,8 @@ export class NodeFS extends FacadeVFS {
   }
 
   deleteDb(fileName: string) {
+    this.onShutdown()
+
     fs.unlinkSync(path.join(this.directory, fileName))
   }
 }